2020-05-19 first

+++ provider/provider-user-api/src/main/java/com/funtl/myshop/plus/provider/api/EmpBaseService.java
@@ -19,6 +19,6 @@ public interface EmpBaseService{
      * @param empBase
      * @return
      */
-    Integer insert(EmpBase empBase);
+    Long insert(EmpBase empBase);
 
 }
